use bigint so we have enough space for a uint32

+++ services/horizon/internal/db2/schema/migrations/55_protocol_19.sql
@@ -1,8 +1,8 @@
 -- +migrate Up
 ALTER TABLE history_transactions ADD ledger_bounds                   int8range;
-ALTER TABLE history_transactions ADD min_account_sequence            integer;
-ALTER TABLE history_transactions ADD min_account_sequence_age        integer;
-ALTER TABLE history_transactions ADD min_account_sequence_ledger_gap integer;
+ALTER TABLE history_transactions ADD min_account_sequence            bigint;
+ALTER TABLE history_transactions ADD min_account_sequence_age        bigint;
+ALTER TABLE history_transactions ADD min_account_sequence_ledger_gap bigint;
 ALTER TABLE history_transactions ADD extra_signers                   character varying(165)[];
 
 -- +migrate Down
